refactor(app): 패키지 분리
비즈니스 로직을 가지는 객체와 UI 관련 객체의 구분을 위하여
각각 domain 패키지, view 패키지로 분리
