+++ src/main/java/com/nextstep/racingcar/domain/Utils.java
@@ -1,12 +1,14 @@
-package com.nextstep.racingcar;
+package com.nextstep.racingcar.domain;
 
 import java.util.Random;
 
 public class Utils {
 
+  private static final int UPPERBOUND = 10;
+
   public static int getRandomInt() {
     Random random = new Random();
-    return random.nextInt(10);
+    return random.nextInt(UPPERBOUND);
   }
 
 }
