+++ src/test/java/com/nextstep/racingcar/CarTest.java
@@ -1,5 +1,7 @@
 package com.nextstep.racingcar;
 
+import com.nextstep.racingcar.domain.Car;
+import com.nextstep.racingcar.domain.CarResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
