+++ src/test/java/com/nextstep/racingcar/RacingcarTest.java
@@ -1,5 +1,7 @@
 package com.nextstep.racingcar;
 
+import com.nextstep.racingcar.domain.Car;
+import com.nextstep.racingcar.domain.Racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
